CI: Don't try to build on forks
The default, free CI runners provided by GitLab are not powerful enough
for OpenEmbedded builds and so we're using a separate runner. This is
selected in the GitLab project settings. However, forked repositories do
not have access to this runner. The result is that builds get stuck on
forked repositories and merge requests never get the magic green tick.

The solution for now is to prevent builds on forks by default. When a
merge request is opened we can push the patches to a `next` branch to
trigger a build. Hopefully this will then be picked up by the merge
request UI.

Signed-off-by: Paul Barker <pbarker@konsulko.com>