linux-stable: Opinionated rework
Previously linux-stable has supported downloading via git or via a
release tarball. Both options are sub-optimal in terms of bandwidth
usage. With git, the entire Linux repo must be downloaded for the first
build. With tarballs, less data is downloaded for the first build but
additional bandwidth and disk space is used for new tarballs every time
a new kernel version is built.

A good compromise is to download one tarball per mainline release (e.g.
5.12) and to apply a patch to get to the desired stable release (e.g.
5.12.9). As the stable patches are small (a couple of MB at most, often
less early on in a release series) this doesn't result in a large
bandwidth and disk space usage over time when following a release
series. A new release tarball is needed when switching kernel release
series but this will happen much less often, especially for users who
want to stay on an LTS release for a few years.

Unfortunately the kernel-yocto class adds additional tasks before
do_patch which modify the source tree. This can lead to the upstream
patches failing to apply so we need to create a new kpatch task which is
guaranteed to run early.

Signed-off-by: Paul Barker <paul@pbarker.dev>