scripts: Rework update-layer
The update-layer script is re-written in Python and now generates the
LINUX_SHA256_TARBALL and LINUX_SHA256_PATCH values needed due to our
updates to linux-stable.inc. We now also automatically update the
linux-mainline recipe as well as each linux-stable version. And as a
bonus the script runs much faster than it used to.

Signed-off-by: Paul Barker <paul@pbarker.dev>