Try to fix automatic branch updates
In GitHub Actions, the checkout action fetches only the latest commit
from the target branch by default. This results in the tracking branches
we want to push to not being present in the working copy and so pushes
are rejected:

    To https://github.com/unnecessary-abstraction/meta-linux-mainline
     ! [rejected]        HEAD -> master (fetch first)
    error: failed to push some refs to 'https://github.com/unnecessary-abstraction/meta-linux-mainline'
    hint: Updates were rejected because the remote contains work that you do
    hint: not have locally. This is usually caused by another repository pushing
    hint: to the same ref. You may want to first integrate the remote changes
    hint: (e.g., 'git pull ...') before pushing again.
    hint: See the 'Note about fast-forwards' in 'git push --help' for details.
    Error: Process completed with exit code 1.

We can try to work around this by forcing the push.

Signed-off-by: Paul Barker <paul@pbarker.dev>