linux-stable, linux-mainline: Use KBUILD_DEFCONFIG as a make target
There is no longer an `rv32_defconfig` file for RISC-V 32-bit support.
Instead there is an `rv32_defconfig` target defined in the Makefile
which combines the RISC-V `defconfig` with the config fragment
`32-bit.config`.

We can't easily support this with the kernel-yocto bbclass as it expects
a real defconfig file which can be copied around. Instead, drop the
reliance on kernel-yocto and treat KBUILD_DEFCONFIG as a make target.

As we're no longer using kernel-yocto, we no longer need to define our
own patch function. Instead, set the download filename for the patch
file so that bitbake will apply it automatically.

Signed-off-by: Paul Barker <paul@pbarker.dev>